+++ web_app/data_source/__init__.py
@@ -1,3 +1,3 @@
 """Init."""
 
-from . import data_source, table_config, utils  # noqa: F401
+from . import data_source, institution_info, table_config, utils  # noqa: F401
